+++ golang/lib/services/docker_container_initializer.go
@@ -38,7 +38,7 @@ type DockerContainerInitializer interface {
 	/*
 		This method is used to declare that the service will need a set of files in order to run. To do this, the developer
 		declares a set of string keys that are meaningful to the developer, and Kurtosis will create one file per key. These newly-createed
-		file objects will then be passed in to the `InitializeMountedFiles` and `GetStartCommand` functions below keyed on the
+		file objects will then be passed in to the `InitializeGeneratedFiles` and `GetStartCommand` functions below keyed on the
 		strings that the developer passed in, so that the developer can initialize the contents of the files as they please.
 		Kurtosis then guarantees that these files will be made available to the service at startup time.
 
@@ -47,21 +47,19 @@ type DockerContainerInitializer interface {
 
 		Returns:
 			A "set" of user-defined key strings identifying the files that the service will need, which is how files will be
-				identified in `InitializeMountedFiles` and `GetStartCommand`
+				identified in `InitializeGeneratedFiles` and `GetStartCommand`
 	*/
-	// TODO Rename "getFilesToGenerate"
-	GetFilesToMount() map[string]bool
+	GetFilesToGenerate() map[string]bool
 
 	/*
-		Initializes the contents of the files that the developer requested in `GetFilesToMount` with whatever
+		Initializes the contents of the files that the developer requested in `GetFilesToGenerate` with whatever
 			contents the developer desires. This will be called before service startup.
 
 		Args:
-			mountedFiles: A mapping of developer_key -> file_pointer, with developer_key corresponding to the keys declares in
-				`GetFilesToMount`
+			filesToGenerate: A mapping of developer_key -> file_pointer, with developer_key corresponding to the keys declares in
+				`GetFilesToGenerate`
 	*/
-	// TODO Rename "initializeFilesToGenerate"
-	InitializeMountedFiles(mountedFiles map[string]*os.File) error
+	InitializeGeneratedFiles(generatedFiles map[string]*os.File) error
 
 	/*
 		Allows the mounting of external files into a service container by mapping files artifacts (defined in your
@@ -78,7 +76,7 @@ type DockerContainerInitializer interface {
 	GetFilesArtifactMountpoints() map[FilesArtifactID]string
 
 	/*
-		Kurtosis mounts the files that the developer requested in `GetFilesToMount` via a Docker volume, but Kurtosis doesn't
+		Kurtosis mounts the files that the developer requested in `GetFilesToGenerate` via a Docker volume, but Kurtosis doesn't
 		know anything about the Docker image backing the service so therefore doesn't know what filepath it can safely mount
 		the volume on. This function uses the developer's knowledge of the Docker image running the service to inform
 		Kurtosis of a filepath where the Docker volume can be safely mounted.
@@ -91,14 +89,10 @@ type DockerContainerInitializer interface {
 	/*
 		Uses the given arguments to build the command that the Docker container running this service will be launched with.
 
-		NOTE: Because the IP address of the container is an implementation detail, any references to the IP address of the
-			container should use the placeholder "SERVICEIP" instead. This will get replaced at launch time with the service's
-			actual IP.
-
 		Args:
-			mountedFileFilepaths: Mapping of developer_key -> initialized_file_filepath where developer_key corresponds to the keys returned
-				in the `GetFilesToMount` function, and initialized_file_filepath is the path *on the Docker container* of where the
-				file has been mounted. The files will have already been initialized via the `InitializeMountedFiles` function.
+			generatedFileFilepaths: Mapping of developer_key -> generated_file_filepath where developer_key corresponds to the keys returned
+				in the `GetFilesToGenerate` function, and generated_file_filepath is the path *on the Docker container* of where the
+				file has been mounted. The files will have already been initialized via the `InitializeGeneratedFiles` function.
 			ipAddr: The IP address of the service being started.
 
 		Returns:
@@ -106,5 +100,5 @@ type DockerContainerInitializer interface {
 				running the service. If this is nil, then no explicit command will be specified and whatever command the Dockerfile
 				specifies will be run instead.
 	*/
-	GetStartCommand(mountedFileFilepaths map[string]string, ipAddr string) ([]string, error)
+	GetStartCommand(generatedFileFilepaths map[string]string, ipAddr string) ([]string, error)
 }